Require Query in Resolver::__construct()
This is an unimplemented feature yet, but behaviors may be query
dependent and the resolver allows for lazy initialization of the
query, which makes things unnecessarily complicated when
implementing this feature.
The resolver is dependent on the query anyway, so it's easier if
the query is already required to build the resolver.