Resolver: Don't *with* relations that are required by expressions
Such are now utilized instead. My reasoning for this simple change
(I had another one in mind, but that turned out to be overkill) is:
Relations that are utilized, are not part of the select list. An
expression is part of the select list of course, but it's an advanced
structure which represents the column's value. How the value is
composed is not of interest, so the relations required by the expression
are not strictly part of the select, but part of the expression. :')