Binary behavior: Return non-resource as is
Otherwise the behavior throws an exception because it requires a
resource in fromDb(). This was created with the background that values
are only retrieved from the database via PDO. However, since behaviors
can be executed explicitly, any value can be passed from any source.
This must not fail if the value is not a resource. For example, we do
this in Icinga DB Web for values read from Redis.