Hydrator: Fix that columns can only be mapped once
Given the two relations `a.b` and `a.b.c` and `a.b.c_id` == `a.b.c.id`,
both paths are qualified to `a_b_c_id`, resulting in only one column
being returned by the datasource. Such a column must be kept and shared
until the last reference is fulfilled.