BelongsToMany: Add support for junction aliases
If you wonder why the junction's actual alias (if there's one)
is not overridden: The only place where this is processed, is
where `resolve()` is called. And that's, at the moment, only
the `Query`. But this uses aliases defined earlier and in
case of `BelongsToMany` relations, this is done when a relation
path is resolved. Which is why you will find a call to
`getThroughAlias()` there.

resolves #147